fix: add campaign empty state component, 1h
