+++ pkg/app/src/components/Account/modules/myCampaignsTab.tsx
@@ -1,6 +1,7 @@
 import React, { useMemo } from 'react'
 
-import { Box } from '@mui/material'
+import { Box, Typography } from '@mui/material'
+import { useTheme } from '@mui/material/styles'
 import { useCurrentAccountState } from 'hooks/useCurrentAccountState'
 import { useTranslation } from 'react-i18next'
 import { Campaign, useCampaignContributorsSubscription, useCampaignSubscription } from 'src/queries'
@@ -8,10 +9,11 @@ import { getAddressFromAccountState } from 'src/utils/accountUtils'
 
 import { CampaignsList } from 'components/CampaignsList/campaignsList'
 import { ContributedCampaignsSection } from 'components/CampaignsSection/ContributedCampaignsSection/contributedCampaignsSection'
+import { CampaignEmptyState } from 'components/CampaignsSection/campaignEmptyState'
 
 export function MyCampaignsTab() {
 	const { t } = useTranslation()
-
+	const theme = useTheme()
 	const accountState = useCurrentAccountState()
 	const { data, loading } = useCampaignSubscription({
 		variables: { address: getAddressFromAccountState(accountState) },
@@ -20,12 +22,18 @@ export function MyCampaignsTab() {
 		useCampaignContributorsSubscription({
 			variables: { address: getAddressFromAccountState(accountState) },
 		})
-
 	const paginatedData = useMemo<Campaign[]>(() => data?.campaign?.slice() as Campaign[], [data])
-
 	return (
 		<Box>
-			<CampaignsList campaigns={paginatedData} loading={loading} title={t('page:campaigns:created_campaigns')} />
+			<Typography marginBottom={2} variant="body2" fontWeight={theme.typography.fontWeightBold}>
+				{t('page:campaigns:created_campaigns')}
+			</Typography>
+			{!loading && paginatedData?.length ? (
+				<CampaignsList campaigns={paginatedData} loading={loading} />
+			) : (
+				<CampaignEmptyState />
+			)}
+
 			<ContributedCampaignsSection data={campaignContributorsData} loading={campaignContributorsLoading} />
 		</Box>
 	)
