+++ pkg/app/src/components/TabPanels/Campaign/overview.tsx
@@ -1,12 +1,9 @@
 import React, { useCallback, useMemo, useState } from 'react'
 
-import { useRouter } from 'next/router'
-
-import { Box, Button, Typography } from '@mui/material'
-import { useTranslation } from 'react-i18next'
 import { Campaign, useCampaignByOrganizationIdSubscription } from 'src/queries'
 
 import { CampaignsList } from 'components/CampaignsList/campaignsList'
+import { CampaignEmptyState } from 'components/CampaignsSection/campaignEmptyState'
 import { CreateCampaignPage } from 'components/TabPanels/Campaign/create'
 
 interface ComponentProps {
@@ -18,11 +15,6 @@ export function CampaignOverview({ organizationId, isAdmin }: ComponentProps) {
 	const { data, loading } = useCampaignByOrganizationIdSubscription({
 		variables: { orgId: organizationId },
 	})
-	const { t } = useTranslation()
-	const { push } = useRouter()
-	const rerouteToCampaigns = useCallback(() => {
-		push('/campaigns')
-	}, [push])
 
 	const paginatedData = useMemo<Campaign[]>(() => data?.campaign?.slice() as Campaign[], [data])
 	const [showCreatePage, setShowCreatePage] = useState<boolean>(false)
@@ -46,25 +38,6 @@ export function CampaignOverview({ organizationId, isAdmin }: ComponentProps) {
 			createCallback={onCreateCallback}
 		/>
 	) : (
-		<Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-			<Box
-				sx={{
-					display: 'flex',
-					flexDirection: 'column',
-					alignContent: 'space-evenly',
-					alignItems: 'center',
-				}}
-			>
-				<Typography variant="subtitle1">{t('page:organisations:no_campaigns')}</Typography>
-				<Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'space-evenly', mt: 2 }}>
-					{isAdmin && (
-						<Button variant="contained" onClick={onCreateCallback}>
-							{t('button:ui:create_campaign')}
-						</Button>
-					)}
-					<Button onClick={rerouteToCampaigns}>{t('button:ui:explore_campaigns')}</Button>
-				</Box>
-			</Box>
-		</Box>
+		<CampaignEmptyState isAdmin={isAdmin} setShowCreatePage={setShowCreatePage} />
 	)
 }
